refactor access control logic by introducing AccessService for centralized authentication and subscription management; replace inline checks in MainChatScreen with service calls; remove redundant code and improve test response handling
